requirements.txt
fastapi[standard]
mysql-connector-python
uvicorn
dotenv
passlib
python-jose[cryptography]